pyproject.toml
[tool.poetry]
name = "everyone-nodong-tgbot"
version = "0.1.0"
description = ""
authors = ["Daehan Kim <really@iamghost.kr>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.13"
python-telegram-bot = "^21.10"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
